fix retention_count bug add more retention_count testcase
